+++ packages/rs/jns42-generator/src/documents/draft_06/mod.rs
