+++ packages/rs/jns42-generator/src/documents/loader.rs
